% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{cartesianExpand}
\alias{cartesianExpand}
\title{expands two data.frames using the Cartesian product}
\usage{
cartesianExpand(x, y)
}
\arguments{
\item{x}{a \code{data.frame}}

\item{y}{a \code{data.frame}}
}
\value{
a \code{data.frame}
}
\description{
takes the cartesian product of two data.frames
}
\examples{
x = data.frame("a" = 1:5, "b" = 6:10)
y = data.frame("z" = letters[1:5], "y" = letters[6:10])
cartesianExpand(x, y)

}
