% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{makeGrid}
\alias{makeGrid}
\title{make a uniform or random grid over some columns of a data.frame}
\usage{
makeGrid(data, vars, n, uniform)
}
\arguments{
\item{data}{a \code{data.frame}}

\item{vars}{integer indices indicating the variables to create the grid for}

\item{n}{two dimensional integer vector giving the resolution of the grid. the first element gives the grid on \code{vars} and the second on the other columns, which are subsampled.}

\item{uniform}{logical, indicates whether a uniform or random grid is to be constructed.}
}
\value{
a \code{data.frame} with \code{n} dimensions.
}
\description{
makes a uniform or random grid over some columns of a data.frame and takes their cartesian product with the other columns
}
\examples{

data <- data.frame(w = seq(0, 1, length.out = 5),
                   x = factor(letters[1:5]),
                   y = ordered(1:5),
                   z = 1:5,
                   r = letters[1:5],
                   stringsAsFactors = FALSE)
makeGrid(data, 1, c(10, 5), TRUE)

}

