% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mi.R
\name{permutationImportance}
\alias{permutationImportance}
\title{computes permutation importance}
\usage{
permutationImportance(data, vars, y, model, nperm = 100L,
  predict.fun = function(object, newdata) predict(object, newdata = newdata),
  loss.fun = function(x, y) defaultLoss(x, y), contrast.fun = function(x, y)
  x - y)
}
\arguments{
\item{data}{a data.frame including the target variable}

\item{vars}{a character vector specifying columns of \code{data} to permute}

\item{y}{a character vector giving the name of the target/outcome variable}

\item{model}{an object with a predict method which returns a vector or matrix. presumably this object represents a model fit.}

\item{nperm}{positive integer giving the number of times to permute the indicated variables (default 100)}

\item{predict.fun}{what function to generate predictions using \code{model}. default is the predict method for \code{model}.}

\item{loss.fun}{what loss function to use to measure prediction errors. default is MSE for ordered predictions and mean misclassification error for unordered prediction errors.}

\item{contrast.fun}{what function to use to contrast the permuted and unpermuted predictions. default is the difference.}
}
\value{
a numeric vector of length one giving the change in prediction error from \code{nperm} permutations of \code{vars}.
}
\description{
computes the change in prediction error from permuting variables
}
\examples{
X = replicate(3, rnorm(100))
y = X \%*\% runif(3)
data = data.frame(X, y)
fit = lm(y ~ -1 + X1 + X2 + X3, data)

permutationImportance(data, "X1", "y", fit)
}

