% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid.R
\name{variableGrid}
\alias{variableGrid}
\title{method to create an evenly spaced grid on a variable}
\usage{
variableGrid(x, length.out)
}
\arguments{
\item{x}{a vector or \code{data.frame} to create a grid on}

\item{length.out}{an integer giving the length of the grid}
}
\value{
an object of the same type as \code{x}, with \code{length.out} unique values
}
\description{
generates and evenly spaced grid given an input vector or \code{data.frame} which has size \code{length.out}
}
\examples{

data = data.frame(
  w = seq(0, 1, length.out = 5),
  x = factor(letters[1:5]),
  y = ordered(1:5),
  z = 1:5
)
variableGrid(data, 3)

}

